refactor(shared): expose trace path builder client-side
