+++ packages/shared/src/index.ts
@@ -18,6 +18,7 @@ export * from "./utils/stringChecks";
 export * from "./utils/objects";
 export * from "./utils/typeChecks";
 export * from "./utils/prompts";
+export * from "./utils/productUrl";
 export * from "./utils/jsonSchemaValidation";
 export * from "./utils/chatml";
 export * from "./utils/math";
